+++ main/src/main/java/ru/practicum/ewm/main/model/Location.java
@@ -15,6 +15,6 @@ public class Location {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Float lat;
-    private Float lon;
+    private Double lat;
+    private Double lon;
 }
